move from_config to new module
